# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/leaderboard": {
      "get": {
        "operationId": "LeaderboardController_getLeaderboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Developer leaderboard retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get developer leaderboard ranked by DQS",
        "tags": [
          "Leaderboard"
        ]
      }
    },
...
    "/debt": {
      "get": {
        "operationId": "DebtController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "repositoryId",
            "required": false,
            "in": "query",
            "description": "Filter by repository ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorId",
            "required": false,
            "in": "query",
            "description": "Filter by author ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "markerType",
            "required": false,
            "in": "query",
            "description": "Filter by marker type",
            "schema": {
              "type": "string",
              "enum": [
                "TODO",
                "FIXME",
                "HACK",
                "XXX"
              ]
            }
          },
          {
            "name": "isResolved",
            "required": false,
            "in": "query",
            "description": "Filter by resolution status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date filter (ISO format)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date filter (ISO format)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filePath",
            "required": false,
            "in": "query",
            "description": "Filter by file path (contains)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of debt items"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all debt items with pagination and filters",
        "tags": [
          "Technical Debt"
        ]
      }
    },
...
    "/debt/{id}": {
      "get": {
        "operationId": "DebtController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Debt item ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Debt item details with related data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Debt item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a debt item by ID",
        "tags": [
          "Technical Debt"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}